refactor(jolt-poly): merge source.rs into multilinear.rs, polish API
- Merge source.rs into multilinear.rs so all three multilinear traits
  (MultilinearEvaluation, MultilinearBinding, MultilinearPoly) and
  RlcSource live in one module
- Make EqPolynomial::evals_serial/evals_parallel pub(crate) — callers
  should use evals() which auto-selects the strategy
- Add Polynomial::into_evals() consuming accessor
- Assert k <= 256 in OneHotPolynomial::new() to match u8 index type
